requirements.txt
asgiref==3.8.1
Django==5.0.6
django-ckeditor-5==0.2.13
django-filter==24.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.3.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
